[DP-10] Обновил код к мерджу дев
