--- DavaiPosmotrim/Resources/Resources.swift
@@ -1,23 +1,0 @@
-//
-//  Resources.swift
-//  DavaiPosmotrim
-//
-//  Created by Sergey Kemenov on 05.05.2024.
-//
-
-import Foundation
-
-enum Resources {
-    enum Mistakes {
-        static let noInternetHeader = "Нет соединения с интернетом"
-        static let noInternetText = "Интернет не бом-бом,\n проверьте соединение"
-        static let noInternetButtonTitle = "повторное подключение".uppercased()
-
-        static let serviceUnavailableHeader = "Сервис временно недоступен"
-        static let serviceUnavailableText = "Дождитесь, пока мы сделаем магию.\n Вжух!"
-
-        static let oldVersionHeader = "Версия приложения устарела"
-        static let oldVersionText = "Сначала сломали, потом починили.\n Скачайте, обновлённую версию."
-        static let oldVersionButtonTitle = "Перейти в магазин".uppercased()
-    }
-}
